finished up some services and .ts files
