+++ social-media-angular/src/app/components/post/post.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
-import Post from 'src/app/models/Post';
+import Post from 'src/app/models/Posts';
 import { AuthService } from 'src/app/services/auth.service';
 import { PostService } from 'src/app/services/post.service';
