--- social-media-angular/src/app/models/Post.ts
@@ -1,17 +1,0 @@
-import User from "./User"
-
-export default class Post {
-    id: number
-    text: string
-    imageUrl: string
-    author: User
-    comments: Post[]
-
-    constructor (id: number, text: string, imageUrl: string, author: User, comments: Post[]) {
-        this.id = id
-        this.text = text
-        this.imageUrl = imageUrl
-        this.author = author
-        this.comments = comments
-    }
-}
