+++ social-media-angular/src/app/models/Posts.ts
@@ -1,0 +1,6 @@
+export interface Post {
+    postId : number;
+    text : string;
+    imageUrl : string;
+    title : string;
+}
