+++ social-media-angular/src/app/services/post.service.ts
@@ -2,7 +2,7 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { environment } from 'src/environments/environment';
-import Post from '../models/Post';
+import { Post } from '../models/Posts';
 
 @Injectable({
   providedIn: 'root'
