--- social-media-angular/src/app/services/posts.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class PostsService {
-
-  constructor() { }
-}
