+++ social-media-angular/src/app/services/user.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http'
-import { User } from './User';
+import { User } from '../models/User';
 
 @Injectable({
   providedIn: 'root'
